+++ feature/main/src/main/java/com/kitching/main/view/order/carditem/CategoryCardItem.kt
@@ -1,15 +1,12 @@
-package com.kitching.main.order.carditem
+package com.kitching.main.view.order.carditem
 
 import com.kitching.main.R
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
