+++ feature/main/src/main/java/com/kitching/main/view/schedule/tab/ScheduleCardItem.kt
@@ -1,4 +1,4 @@
-package com.kitching.main.schedule.tab
+package com.kitching.main.view.schedule.tab
 
 import com.kitching.main.R
 import androidx.compose.foundation.border
