+++ feature/main/src/main/java/com/kitching/main/view/schedule/tab/ScheduleListScreen.kt
@@ -1,9 +1,8 @@
-package com.kitching.main.schedule.tab
+package com.kitching.main.view.schedule.tab
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import com.kitching.main.R
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
